refactor: move child components folders into parent component folders
