+++ src/Round/Round.js
@@ -4,8 +4,8 @@ import styles from './Round.module.css';
 import { useState, useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 
-import RoundHeader from '../RoundHeader/RoundHeader';
-import AlbumList from '../AlbumList/AlbumList';
+import RoundHeader from './RoundHeader/RoundHeader';
+import AlbumList from './AlbumList/AlbumList';
 
 function Round() {
